fw: ov7670 camra colors corrected

+++ firmware/ov7670_camera/Python/reciever.py
@@ -106,7 +106,8 @@ print(frame[:32].hex())
 # RGB565:
 # RRRRRGGGGGGBBBBB
 
-data = np.frombuffer(frame, dtype=">u2")
+# data = np.frombuffer(frame, dtype=">u2")
+data = np.frombuffer(frame, dtype="<u2")
 
 r = (data >> 11) & 0x1F
 g = (data >> 5) & 0x3F
